apps/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==3.6.4.0
bitarray==2.9.3
boto3==1.26.137
botocore==1.29.165
cachetools==5.5.2
celery==5.2.7
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.2.3
cryptography==41.0.3
cytoolz==1.0.1
Django==4.2
django-cors-headers==4.0.0
django-extensions==3.2.0
django-filter==23.1
django-guardian==2.4.0
django-ratelimit==4.0.0
django-redis==5.2.0
django-storages==1.13.2
django-tenant-schemas==1.10.0
django-tenants==3.8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dotenv==0.9.9
drf-spectacular==0.27.0
eth-account==0.8.0
eth-hash==0.7.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.2
eth-utils==2.3.2
eth_abi==5.0.1
frozenlist==1.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.4.4
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.62.3
gunicorn==20.1.0
hexbytes==0.3.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
lru-dict==1.3.0
multidict==6.6.4
openai==0.27.8
ordered-set==4.1.0
packaging==25.0
parsimonious==0.9.0
Pillow==9.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psycopg2-binary==2.9.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
PyJWT==2.10.1
pytesseract==0.3.10
pytest==7.3.1
pytest-django==4.5.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==4.5.4
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rlp==3.0.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.6.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
toolz==1.0.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
web3==6.0.0
websockets==15.0.1
whitenoise==6.4.0
yarl==1.20.1
